broke out the main file even more
