+++ notes.txt
@@ -30,4 +30,4 @@ Figure out how to make container prod and run on cloud run
 
 
 docker build -t gcr.io/kjtmckenzie-home-fs/kjtmckenzie-home-fs:latest .              
-docker run -e "PORT=8080, " -p 8080:8080 gcr.io/kjtmckenzie-home-fs/kjtmckenzie-home-fs
+docker run -e "PORT=8080" -p 8080:8080 gcr.io/kjtmckenzie-home-fs/kjtmckenzie-home-fs
